+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.1.0 - 2018-04-12
+
+### Added
+
+- Added `replace` method to `Step`, `Scenario`, `TableCell`, `TableRow`, `DocString` and `DataTable`.
+- Added `toScenario` method to `ScenarioOutline`.
+
 ## 1.0.1 - 2018-03-23
 
 ### Fixed